Updated project settings. Remove warnings regarding implicit conversions and awakeFromNib.

+++ Samples/JotFormTest/Classes/Controller/CreateFormViewController.m
@@ -78,7 +78,7 @@
         [SVProgressHUD dismiss];
         
         if ( result != nil ) {
-            int responseCode = [[result objectForKey:@"responseCode"] integerValue];
+            NSInteger responseCode = [[result objectForKey:@"responseCode"] integerValue];
             
             if ( responseCode == 200 || responseCode == 206 ) {
                 
@@ -92,8 +92,7 @@
         [SVProgressHUD dismiss];
         
         if ( error != nil ) {
-            
-            int responseCode = [[error objectForKey:@"response"] integerValue];
+            NSInteger responseCode = [[error objectForKey:@"response"] integerValue];
             
             if ( responseCode == 401 ) {
